Tests/App: refactor examples - structure (#15770)
* rename _examples dir

* refactor

* clean

* path

* add inits

* skip

* e2e

* azure

* e2e

* rev

* unify single depth for ignore docs req.

* group

(cherry picked from commit 59fa32080d5256221aa36da56a735f28ed8c06ef)
