+++ .azure/app-cloud-e2e.yml
@@ -27,16 +27,15 @@ pr:
       - "release/*"
   paths:
     include:
-      - ".azure/app-cloud-e2e.yml"
-      - "requirements/app/**"
-      - "src/lightning_app/**"
-      - "tests/tests_app/**"
-      - "examples/app_*/**"  # some tests_app tests call examples files
-      - "tests/tests_app_examples/**"
-      - "setup.py"
       - ".actions/**"
+      - ".azure/app-cloud-e2e.yml"
+      - "src/lightning_app/**"
+      - "examples/app_*/**"  # some tests_app tests call examples files
+      - "requirements/app/**"
+      - "tests/tests_examples_app/**"
+      - "setup.py"
     exclude:
-      - "requirements/app/docs.txt"
+      - "requirements/*/docs.txt"
       - "*.md"
       - "**/*.md"
 
@@ -59,30 +58,42 @@ jobs:
       matrix:
         'App: v0_app':
           name: "v0_app"
+          dir: "public"
         'App: boring_app':
           name: "boring_app"
+          dir: "public"
         'App: template_streamlit_ui':
           name: "template_streamlit_ui"
+          dir: "public"
         'App: template_react_ui':
           name: "template_react_ui"
+          dir: "public"
         # 'App: template_jupyterlab':  # TODO: clarify where these files lives
         #   name: "template_jupyterlab"
-        'App: idle_timeout':
-          name: "idle_timeout"
-        'App: collect_failures':
-          name: "collect_failures"
-        'App: custom_work_dependencies':
-          name: "custom_work_dependencies"
         'App: installation_commands_app':
           name: "installation_commands_app"
+          dir: "public"
         'App: drive':
           name: "drive"
+          dir: "public"
         'App: payload':
           name: "payload"
+          dir: "public"
         'App: commands_and_api':
           name: "commands_and_api"
+          dir: "public"
         'App: quick_start':
           name: "quick_start"
+          dir: "public"
+        'App: idle_timeout':
+          name: "idle_timeout"
+          dir: "local"
+        'App: collect_failures':
+          name: "collect_failures"
+          dir: "local"
+        'App: custom_work_dependencies':
+          name: "custom_work_dependencies"
+          dir: "local"
     timeoutInMinutes: "20"
     cancelTimeoutInMinutes: "2"
     # values: https://docs.microsoft.com/en-us/azure/devops/pipelines/process/phases?view=azure-devops&tabs=yaml#workspace
@@ -124,7 +135,7 @@ jobs:
     # - bash: |
     #     rm -rf examples/app_template_jupyterlab || true
     #     git clone https://github.com/Lightning-AI/LAI-lightning-template-jupyterlab-App examples/app_template_jupyterlab
-    #     cp examples/app_template_jupyterlab/tests/test_template_jupyterlab.py tests/tests_app_examples/test_template_jupyterlab.py
+    #     cp examples/app_template_jupyterlab/tests/test_template_jupyterlab.py tests/tests_examples_app/test_template_jupyterlab.py
     #   condition: eq(variables['name'], 'template_jupyterlab')
     #   displayName: 'Clone Template Jupyter Lab Repo'
 
@@ -152,8 +163,9 @@ jobs:
     - bash: |
         mkdir -p ${VIDEO_LOCATION}
         ls -l examples/${TEST_APP_NAME}
-        ls -l tests/tests_app_examples
-        python -m pytest tests/tests_app_examples/test_${TEST_APP_NAME}.py::test_${TEST_APP_NAME}_example_cloud --timeout=1200 --capture=no -v --color=yes
+        ls -l tests/tests_examples_app/public
+        python -m pytest tests/tests_examples_app/${TEST_APP_FOLDER}/test_${TEST_APP_NAME}.py::test_${TEST_APP_NAME}_example_cloud \
+          --timeout=1200 --capture=no -v --color=yes
       env:
         HEADLESS: '1'
         PACKAGE_LIGHTNING: '1'
@@ -161,6 +173,7 @@ jobs:
         VIDEO_LOCATION: $(video_artifact_dir)
         PR_NUMBER: $(local_id)
         TEST_APP_NAME: $(name)
+        TEST_APP_FOLDER: $(dir)
         HAR_LOCATION: './artifacts/hars'
         SLOW_MO: '50'
         #LAI_USER: $(LAI_USER)  # for STAGING
@@ -191,6 +204,7 @@ jobs:
         LIGHTNING_CLOUD_URL: $(LIGHTNING_CLOUD_URL_PROD)
         PR_NUMBER: $(local_id)
         TEST_APP_NAME: $(name)
+        TEST_APP_FOLDER: $(dir)
         # GRID_USER_ID: $(LIGHTNING_USER_ID)  # TODO: clarify the meaning
         # GRID_USER_KEY: $(LIGHTNING_API_KEY)  # TODO: clarify the meaning
         # GRID_URL: $(LIGHTNING_CLOUD_URL)
