+++ .azure/gpu-tests-lite.yml
@@ -20,6 +20,7 @@ pr:
       - "release/*"
   paths:
     include:
+      - ".actions/**"
       - ".azure/gpu-tests-lite.yml"
       - "examples/lite/**"
       - "examples/run_lite_examples.sh"
@@ -29,7 +30,6 @@ pr:
       - "src/lightning_lite/**"
       - "tests/tests_lite/**"
       - "setup.cfg"  # includes pytest config
-      - ".actions/**"
     exclude:
       - "requirements/lite/docs.txt"
       - "*.md"
