+++ .azure/gpu-tests-pytorch.yml
@@ -20,6 +20,7 @@ pr:
       - "release/*"
   paths:
     include:
+      - ".actions/**"
       - ".azure/gpu-tests-pytorch.yml"
       - "examples/run_ddp_examples.sh"
       - "examples/convert_from_pt_to_pl/**"
@@ -36,9 +37,8 @@ pr:
       - "setup.cfg"  # includes pytest config
       - "requirements/lite/**"
       - "src/lightning_lite/**"
-      - ".actions/**"
     exclude:
-      - "requirements/**/docs.txt"
+      - "requirements/*/docs.txt"
       - "*.md"
       - "**/*.md"
