+++ .azure/hpu-tests.yml
@@ -17,6 +17,7 @@ pr:
       - "release/*"
   paths:
     include:
+      - ".actions/**"
       - ".azure/hpu-tests.yml"
       - "examples/pl_hpu/mnist_sample.py"
       - "requirements/lite/**"
@@ -25,9 +26,8 @@ pr:
       - "src/pytorch_lightning/**"
       - "tests/tests_pytorch/**"
       - "setup.cfg"  # includes pytest config
-      - ".actions/**"
     exclude:
-      - "requirements/**/docs.txt"
+      - "requirements/*/docs.txt"
       - "*.md"
       - "**/*.md"
