+++ .azure/ipu-tests.yml
@@ -15,6 +15,7 @@ pr:
       - "release/*"
   paths:
     include:
+      - ".actions/**"
       - ".azure/ipu-tests.yml"
       - "requirements/lite/**"
       - "src/lightning_lite/**"
@@ -22,9 +23,8 @@ pr:
       - "src/pytorch_lightning/**"
       - "tests/tests_pytorch/**"
       - "setup.cfg"  # includes pytest config
-      - ".actions/**"
     exclude:
-      - "requirements/**/docs.txt"
+      - "requirements/*/docs.txt"
       - "*.md"
       - "**/*.md"
