--- tests/tests_app_examples/test_app_dag.py
@@ -1,21 +1,0 @@
-import os
-from time import sleep
-
-import pytest
-from tests_app import _PROJECT_ROOT
-
-from lightning_app.testing.testing import run_app_in_cloud
-
-
-@pytest.mark.cloud
-def test_app_dag_example_cloud() -> None:
-    with run_app_in_cloud(os.path.join(_PROJECT_ROOT, "examples/app_dag")) as (_, _, fetch_logs, _):
-
-        launch_log, finish_log = False, False
-        while not (launch_log and finish_log):
-            for log in fetch_logs(["flow"]):
-                if "Launching a new DAG" in log:
-                    launch_log = True
-                elif "Finished training and evaluating" in log:
-                    finish_log = True
-            sleep(1)
