+++ tests/tests_examples_app/apps/idle_timeout/__init__.py
@@ -1,0 +1,1 @@
+# TODO: check the tests to be able to run without this init file
