+++ tests/tests_examples_app/local/__init__.py
@@ -1,0 +1,5 @@
+import os
+
+from lightning_app import _PROJECT_ROOT
+
+_PATH_APPS = os.path.join(_PROJECT_ROOT, "tests", "tests_examples_app", "apps")
