+++ tests/tests_examples_app/public/__init__.py
@@ -1,0 +1,5 @@
+import os
+
+from lightning_app import _PROJECT_ROOT
+
+_PATH_EXAMPLES = os.path.join(_PROJECT_ROOT, "examples")
